Add inpage-toolbar-ui add-on example
This example shows how an add-on can integrate an UI into a webpage.

The UI resources will be packaged in the add-on itself and exposed to
the webpage by injecting an iframe which points to an "web accessible
resource" add-on URL.

This pattern is useful to any add-on which needs to integrate an UI into
the webpage (e.g. browsing enhancement toolbars, in page devtools
panels,  adblockers/anti-tracking in-page alerts) and it works on
Chromium-based browsers and Firefox Desktop >= 46.0 (enabled by Bug
1214658).