remove inpage-toolbar-ui example
Based on the discussion in #74, we decided to remove this example
to prevent it to seem as the suggested way to expose
"security/privacy" sensible UI elements to a potentially malicious page.

Even if there are probably reasonable scenarios where this feature
can be used, it is currently perceived as a replacement for a
browser toolbar UI element, which is not.