Update the `commands` sample to have 2 commands
Before the change, the sample showed how to add one command, and how to
update it, reset it, and respond to it. But usually the extension
developer needs to configure more than one command, and the code was not
readily usable for such a need.

After this patch, the developer can easily see what are the different
places they need to make conditional decisions, to add a new command to
the extension.

This patch also includes some cosmetic changes, like wrapping the long
lines of the README.md, as intended by the authors of Markdown syntax.
If desirable, I can put such cosmetic changes in a separate commit.