Correct userScripts / userScripts_legacy index (#579)
The MV2-specific userScripts example is currently associated with the
"userScripts.register" API. While this is the correct API name, the
MDN documentation is actually at `userScripts_legacy/register`.

To make sure that the example is rendered in the right location,
change "userScripts" to "userScripts_legacy".

... and for completeness also add `runtime.onMessage` to the MV3
example, since that example uses this API.